Reducing Line of code to increase Readability
Storing direct response from database

Note:
In general, if the API response already provides the data in the format required by your application (in this case, the BookModel interface), you can directly use responseJSON._embedded.books and avoid the extra overhead of the loop and manual transformation. However, if there is a need to modify or adapt the data before using it, then you would opt for the first code snippet with the loop and transformation logic.